requirements.txt
keras==2.11.0
matplotlib==3.6.3
mediapipe==0.9.0.1
numpy==1.24.2
opencv_python_headless==4.7.0.72
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.10.1
streamlit==1.19.0
tensorflow==2.11.0